requirements.txt
snowflake-connector-python
snowflake-connector-python[pandas]
requests
tenacity
python-dotenv
pandas